fix(test): rename internal/root/root to internal/root/rootroot; use HandlerFunc
