+++ internal/root/rootroot/handler.go
@@ -1,4 +1,4 @@
-package root
+package rootroot
 
 import (
 	"net/http"
